Start parser work.
Downloaded two parser libraries:

https://github.com/yhirose/cpp-peglib : Commit a517ce662259043ee1b8fd8899a319e669acc929
https://github.com/taocpp/PEGTL : Release 2.4.0

They're currently included directly with the repository (luckily the
files weren't too big). We'll change that later.

Starting to work with some examples for the two parsers. Right now,
PEGTL uses advanced features of C++ and our team is rusty with C++. I'm
not sure it's a good idea to try and build with something that we don't
really understand, esepcially when those language features are
apparently horrendous to debug. However, PEGTL is probably faster at
run-time since a lot of the work of compiling a function to build a
parser is done at compile-time instead of run-time.

Changes to be committed:
      modified:   Makefile
      new file:   include/peglib.h
      new file:   parser.cpp
      new file:   tao-parser-example.cpp
      new file:   include/tao/