Implement +(Mem, Mem), +(In, In).
Start implementing RMS instructions. Must implement:

- *(Mem, In) : Multiplying a memory location and an input.
- +(Mem, Mem) : Adding two memory locations together (DONE).

I'm also trying to implement the following, though they're not
officially part of the language.

- +(In, In)

Also created a base class for inputs and memory locations that allows
the addition of two RMS objects without knowing their types at compile
time. This was necessary since an RMS program's structure is not known
at compile-time. Though it could be. There's a way to do it, based on
the work at <https://github.com/taocpp/PEGTL>. Our applications are not
so different.