Fix bug in ZZToBits and Input constructor
- ZZToBytes requires you to allocate space for the bytes. It does not
allocate space for you. So I had to change ZZToBits to reflect this.
- In ZZToBytes and Input constructor, I initialzed my bit vectors with
an initial size. This means putting in a bunch of zero entries to start,
so my vectors  were double the size that they should be, with the first
half being all 0's and the latter half being the correct contents.

Changes to be committed:
      modified:   paillier/data.cpp