Fix GenPrimePair, create encryptBits
GenPrimePair's stop condition was wrong for generating a new prime.
Changed that.

Also created the encryptBits function, which returns a vector of
encryptions of bits of the secret key. This is intended for use with the
RMS operations, so that it is easy to create inputs, so long as you have
a Paillier object.

Changes to be committed:
      modified:   paillier/paillier.cpp
      modified:   paillier/paillier.h